Multinomial Naive Bayes Algorithm is used to detect Spam Messages
Multinomial Naive Bayes Algorithm is similar to Naive Bayes Algorithm but is used on discrete valued input (eg: values that can be finitely counted). Here spam messages are divided into words which can be arranged as discrete values (1- word is there in sentence, 0-word is not there in sentence). This algorithm is used to find if a message is spam or ham.