Implementing Naive Bayes Algorithm on Sentences (Count Vectorizer)
This code shows how Naive Bayes Algorithm is implemented on sentences to give an outcome value. The sentence is broken down into words and then into numerical values. This pre-processing can be done manually or using a Count Vectorizer. This numerical data is then used to calculate the probability of an outcome value given the input.