Combining Gaussian Naive Bayes and normal Naive Bayes Algorithm for mixed data
This code shows Implementation of Gaussain Naive Bayes Algorithm for continuous valued input attributes and implementation of normal Naive Bayes Algorithm for categorical valued input attributes of a sample/instance to predict its output. This method is used in calculating outcome when input consists a mix of categorical and continuous values.